+++ CHANGELOG.md
@@ -64,6 +64,7 @@
     + `ESPathForBundleResource()`, `ESPathForMainBundleResource()`
     + `ESTouchDirectory()`, `ESTouchDirectoryAtFilePath`, `ESTouchDirectoryAtFileURL()`: use `-[NSFileManager createDirectoryAtPath:]` or `-[NSFileManager createDirectoryAtURL:]` instead before writing to the filesystem
     + `NSNumberFromString()`
+    + `-[NSObject es_description]`
     + `NSObject (ESAssociatedObjectHelper)` category
     + `NSRegularExpression (ESAdditions)` category
     + `UIAlertView+ESBlock`, `UIActionSheet+ESBlock`